+++ Framework/Cache/ICache.cs
@@ -27,7 +27,7 @@ namespace Cache
         /// <param name="func">委托方法</param>
         /// <param name="expiration">缓存有效期（单位：秒）</param>
         /// <returns></returns>
-        T GetOrSet<T>(string key, Func<T> func, uint expiration = 0);
+        T GetOrSet<T>(string key, Func<T> func, ulong expiration = 0);
 
         /// <summary>
         /// 移除缓存
@@ -41,7 +41,7 @@ namespace Cache
         /// <param name="key">缓存key</param>
         /// <param name="value">缓存值</param>
         /// <param name="second">缓存有效期（单位：秒）</param>
-        void Set(string key, string value, uint expiration = 0);
+        void Set(string key, string value, ulong expiration = 0);
 
         /// <summary>
         /// 设置缓存
@@ -50,6 +50,6 @@ namespace Cache
         /// <param name="key">缓存key</param>
         /// <param name="value">缓存值</param>
         /// <param name="expiration">缓存有效期（单位：秒）</param>
-        void Set<T>(string key, T value, uint expiration = 0);
+        void Set<T>(string key, T value, ulong expiration = 0);
     }
 }
