+++ Framework/Cache/Local/LocalCache.cs
@@ -40,7 +40,7 @@ namespace Cache.Local
         /// <param name="key">缓存key</param>
         /// <param name="value">缓存值</param>
         /// <param name="expiration">缓存有效期（单位：秒）</param>
-        public override void Set<T>(string key, T value, uint expiration = 0)
+        public override void Set<T>(string key, T value, ulong expiration = 0)
         {
             if (expiration > 0)
             {
